Different Secret keys to save different information
Security is important;
Added different keys for different encryptation values {games_won,
games_finished, game_state}, so they can not be replaced (i.e. set
number of games played as number of games won)

It could have been changed by adding another identifier to the encrypted
string and just read it when decrypting, but I wanted to try this way.

Now the only way to hack this is to know the secret keys or users
trading settings file, should I be saying this?

bye