Now sign-out issue is being taken care by GameHelper.Java 2
comment from:
https://github.com/playgameservices/android-basic-samples/issues/121

“When you start a GPGS activity, you have to start it with
startActivityForResult. If the user clicks on sign-out, the
GoogleApiClient is NOT disconnected - it's just in a "signed out"
state. The onActivityResult here will return a status
RECONNECT_REQUIRED. At this point, the developer should reconnect their
client, and life goes on as expected.”

So I just call disconnect function if we get the  RECONNECT_REQUIRED
responseCode