Skip splash screen if the main activity has been initialized.
If the main activity was initialized and then the player pressed home
or jumped to another app, the next time the player clicks on a push
notification or the app icon, we will skip the splash screen.

Also I added a new intent flag: Intent.FLAG_ACTIVITY_CLEAR_TOP, to
prevent creating many times the main activity.