Fix problem of saved game not being loaded on next start of game state
- There is a huge problem with some android devices taking long time to
load a save game from local preferences I have to do more tests with
those devices to track down why is taking so long.

the not so pretty fix is to just create a dummy default board at the
start of each game, so if the game is taking too much time loading the
save, we have a default instance to fall back, AND we don’t tracking it
as new board unless there is no save to load.

- Removed some sysouts prints

- GcmReceiver threw an exception on first run when trying to
setResultCode, nothing important since game was able to keep working.